Add audit txma events for issue status list entry. lambda (#45)
## JIRA Ticket

<!-- Update as appropriate with the ID of the JIRA ticket for which the
PR has been raised. -->

[DCMAW-12604](https://govukverify.atlassian.net/browse/DCMAW-12604)

## Description of Changes

Add audit events (txma events) that are sent to the sqs queue
Added jwtVerify to enable public key verification (hardcoded but can be
updated once we set up our key store)
Add tests for above

## Evidence
![Screenshot 2025-05-20 at 11 10
41](https://github.com/user-attachments/assets/36ced3fc-73b4-4033-8b59-fb8b661ecbbc)
![Screenshot 2025-05-20 at 13 47
03](https://github.com/user-attachments/assets/ebd745a4-99f9-4380-b639-a789c5796eb7)

## Documentation

<!-- 
Links to documentation that has been created or updated to capture the
changes.

Major changes to the way the service is run and tested, or to the
structure of this repository, should be
captured in the README.

Changes to API endpoints/contracts should be captured in the appropriate
OpenAPI spec.

Changes to the following should be captured in the appropriate Mobile
Backend reference documentation on Confluence

(https://govukverify.atlassian.net/wiki/spaces/DCMAW/pages/3861905465/Mobile+Backend+-+Reference):
  - Endpoints
  - Environment variables/configuration
  - Dependencies
-->

---

<details>
<summary><h2>Guidance for Reviewers</h2></summary>

### Functionality
* Have all of the ticket requirements and acceptance criteria been met?
* Will this code function as expected in different environments?

### Testing

* Is every component covered by high quality unit tests?
* Have API/end-to-end tests been updated to cover key new behaviour?
* Have infra tests been added if needed for changes to our templates?

### Logging

* Are the changes supported by sufficient logging to help support, debug
and analyse the behaviour of the service?
* Is anything being logged that shouldn't be? (PII, redundant
information)

### Code Quality

* Does the code conform to our [values, principles and
practices](https://govukverify.atlassian.net/wiki/spaces/DCMAW/pages/3910271319/Values+Principles+and+Practices)
for high quality code?

### Documentation

* Has any relevant reference documentation been updated?
* Would any further documentation of the changes make the service easier
to understand and maintain?

</details>


[DCMAW-12604]:
https://govukverify.atlassian.net/browse/DCMAW-12604?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ