BE | Status List | IssueStatusListEntry | Add item to StatusListTable (#48)
## JIRA Ticket

[[DCMAW-11869](https://govukverify.atlassian.net/browse/DCMAW-11869)]

## Description of Changes

Added in logic to take messages of the correct sqs queue
Added in logic and tests to add a credential to the status list table.
Added in validation to the next available index - to see if it exists
already and if it does it polls again.

## Evidence
<img width="899" alt="Screenshot 2025-05-27 at 16 33 23"
src="https://github.com/user-attachments/assets/6a609dfd-80d2-41f4-97bf-5bb369877f30"
/>

## Documentation

<!-- 
Links to documentation that has been created or updated to capture the
changes.

Major changes to the way the service is run and tested, or to the
structure of this repository, should be
captured in the README.

Changes to API endpoints/contracts should be captured in the appropriate
OpenAPI spec.

Changes to the following should be captured in the appropriate Mobile
Backend reference documentation on Confluence

(https://govukverify.atlassian.net/wiki/spaces/DCMAW/pages/3861905465/Mobile+Backend+-+Reference):
  - Endpoints
  - Environment variables/configuration
  - Dependencies
-->

---

<details>
<summary><h2>Guidance for Reviewers</h2></summary>

### Functionality
* Have all of the ticket requirements and acceptance criteria been met?
* Will this code function as expected in different environments?

### Testing

* Is every component covered by high quality unit tests?
* Have API/end-to-end tests been updated to cover key new behaviour?
* Have infra tests been added if needed for changes to our templates?

### Logging

* Are the changes supported by sufficient logging to help support, debug
and analyse the behaviour of the service?
* Is anything being logged that shouldn't be? (PII, redundant
information)

### Code Quality

* Does the code conform to our [values, principles and
practices](https://govukverify.atlassian.net/wiki/spaces/DCMAW/pages/3910271319/Values+Principles+and+Practices)
for high quality code?

### Documentation

* Has any relevant reference documentation been updated?
* Would any further documentation of the changes make the service easier
to understand and maintain?

</details>


[DCMAW-11869]:
https://govukverify.atlassian.net/browse/DCMAW-11869?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ