DCMAW-13149: CloudWatch alerts for FAS errors (#42)
## JIRA Ticket

[DCMAW-13149](https://govukverify.atlassian.net/browse/DCMAW-13149)

## Description of Changes

Added a new CloudWatch alarm that monitors the execution status of the
FAS lambda, and should the lambda return in error, the alert transitions
to ALARM mode, which will cause the newly created state machine to
invoke the FAS lambda again.

## Evidence


![image](https://github.com/user-attachments/assets/e217ea27-1428-4a89-b773-cccf8d5238dd)
The screenshot shows repeated failed runs of the FAS lambda (pre 12:24,
simulated/forced error), a successful run that filled up the Bitstring
SQS queue (12:24, root of error removed) and the subsequent 5 successful
runs that are documented below (post 12:24). All runs were invoked by
the state machine presented in the screenshot below.


![image](https://github.com/user-attachments/assets/bec79502-dc1d-48de-90a6-be649cca3152)


## Documentation

A new tiny lambda was added to read the status of the CloudWatch alert
that monitors the errors that may occur when the FAS lambda is running.
A state machine checks the output of the new lambda and decides whether
to invoke the FAS lambda again (in case of error) or do nothing in case
no errors were thrown. The state machine checks the status of the alert
every minute.

Note: in the case of a state transition from ALARM to OK of the Error
alarm, the state machine will invoke the FAS an additional 5 times
because of the delay the CloudWatch metric inherently has before
updating its state (approximately 5 minutes) and because the state
machine checks for updates every minute.

---

<details>
<summary><h2>Guidance for Reviewers</h2></summary>

### Functionality
* Have all of the ticket requirements and acceptance criteria been met?
* Will this code function as expected in different environments?

### Testing

* Is every component covered by high quality unit tests?
* Have API/end-to-end tests been updated to cover key new behaviour?
* Have infra tests been added if needed for changes to our templates?

### Logging

* Are the changes supported by sufficient logging to help support, debug
and analyse the behaviour of the service?
* Is anything being logged that shouldn't be? (PII, redundant
information)

### Code Quality

* Does the code conform to our [values, principles and
practices](https://govukverify.atlassian.net/wiki/spaces/DCMAW/pages/3910271319/Values+Principles+and+Practices)
for high quality code?

### Documentation

* Has any relevant reference documentation been updated?
* Would any further documentation of the changes make the service easier
to understand and maintain?

</details>


[DCMAW-13149]:
https://govukverify.atlassian.net/browse/DCMAW-13149?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Amrit Kandola <amrit.kandola@digital.cabinet-office.gov.uk>