DCMAW-11956: Adding Revoke lambda alarms (#50)
## JIRA Ticket

[DCMAW-11956](https://govukverify.atlassian.net/browse/DCMAW-11956)

## Description of Changes

Added throughput, latency, concurrent, 4xx (low and high), and 5xx (low
and high) alarms as well as accompanying sns topics as well as infra
test. Infra test to ensure all critical alarms have runbooks has NOT
been included as not alarms are currently critical. This test will be
added when we are closer to moving to production and runbooks are
created.

## Evidence

<!-- Evidence that the changes are working as expected, e.g. screenshots
from the dev environment -->

## Documentation

<!-- 
Links to documentation that has been created or updated to capture the
changes.

Major changes to the way the service is run and tested, or to the
structure of this repository, should be
captured in the README.

Changes to API endpoints/contracts should be captured in the appropriate
OpenAPI spec.

Changes to the following should be captured in the appropriate Mobile
Backend reference documentation on Confluence

(https://govukverify.atlassian.net/wiki/spaces/DCMAW/pages/3861905465/Mobile+Backend+-+Reference):
  - Endpoints
  - Environment variables/configuration
  - Dependencies
-->

---

<details>
<summary><h2>Guidance for Reviewers</h2></summary>

### Functionality
* Have all of the ticket requirements and acceptance criteria been met?
* Will this code function as expected in different environments?

### Testing

* Is every component covered by high quality unit tests?
* Have API/end-to-end tests been updated to cover key new behaviour?
* Have infra tests been added if needed for changes to our templates?

### Logging

* Are the changes supported by sufficient logging to help support, debug
and analyse the behaviour of the service?
* Is anything being logged that shouldn't be? (PII, redundant
information)

### Code Quality

* Does the code conform to our [values, principles and
practices](https://govukverify.atlassian.net/wiki/spaces/DCMAW/pages/3910271319/Values+Principles+and+Practices)
for high quality code?

### Documentation

* Has any relevant reference documentation been updated?
* Would any further documentation of the changes make the service easier
to understand and maintain?

</details>


[DCMAW-11956]:
https://govukverify.atlassian.net/browse/DCMAW-11956?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ