Dcmaw 11959 revoke consult client registry (#59)
## JIRA Ticket

<!-- Update as appropriate with the ID of the JIRA ticket for which the
PR has been raised. -->

[DCMAW-11959](https://govukverify.atlassian.net/browse/DCMAW-11959)

## Description of Changes

I have added error handling for the revoke lambda,
Added logic to consult the client registry etc.
Added logic to make sure that it consults the db to check that the
clientId that issued the credential matches the one thats on the request
for revoke.
Added testing for all of the above
Added some extra uri functionality on the issue lambda (small change but
good for future)

## Evidence

No screenshots - as there is no data that is signed by any of the
clients from the wellknown endpoint

## Documentation

<!-- 
Links to documentation that has been created or updated to capture the
changes.

Major changes to the way the service is run and tested, or to the
structure of this repository, should be
captured in the README.

Changes to API endpoints/contracts should be captured in the appropriate
OpenAPI spec.

Changes to the following should be captured in the appropriate Mobile
Backend reference documentation on Confluence

(https://govukverify.atlassian.net/wiki/spaces/DCMAW/pages/3861905465/Mobile+Backend+-+Reference):
  - Endpoints
  - Environment variables/configuration
  - Dependencies
-->

---

<details>
<summary><h2>Guidance for Reviewers</h2></summary>

### Functionality
* Have all of the ticket requirements and acceptance criteria been met?
* Will this code function as expected in different environments?

### Testing

* Is every component covered by high quality unit tests?
* Have API/end-to-end tests been updated to cover key new behaviour?
* Have infra tests been added if needed for changes to our templates?

### Logging

* Are the changes supported by sufficient logging to help support, debug
and analyse the behaviour of the service?
* Is anything being logged that shouldn't be? (PII, redundant
information)

### Code Quality

* Does the code conform to our [values, principles and
practices](https://govukverify.atlassian.net/wiki/spaces/DCMAW/pages/3910271319/Values+Principles+and+Practices)
for high quality code?

### Documentation

* Has any relevant reference documentation been updated?
* Would any further documentation of the changes make the service easier
to understand and maintain?

</details>


[DCMAW-11959]:
https://govukverify.atlassian.net/browse/DCMAW-11959?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ